requirements.txt
# Python 基本ツール
ipykernel
jupyter

# LangChain 本体と関連パッケージ
langchain==1.1.2
langchain-community==0.4.1
langchain-openai==1.1.0
langchain-core==1.1.0
langchain-chroma==1.0.0
pypdf

# .envなどを使う場合
python-dotenv

# その他、必要に応じて
tiktoken
requests
python-dotenv==1.2.1
streamlit==1.52.2
fastapi
uvicorn[standard]
pydantic
